requirements.txt
django
django-suit-dashboard
django-suit
psutil
